Only download test DB for CLI integration tests
The test database is not required for activated extension tests, so we
only need to download the test database for CLI integration tests.
