+++ extensions/ql-vscode/test/vscode-tests/activated-extension/jest.setup.ts
@@ -1,1 +1,12 @@
-import "../jest.activated-extension.setup";
+import {
+  beforeAllAction,
+  beforeEachAction,
+} from "../jest.activated-extension.setup";
+
+beforeAll(async () => {
+  await beforeAllAction();
+});
+
+beforeEach(async () => {
+  await beforeEachAction();
+});
